+++ pixi.lock
@@ -14706,9 +14706,10 @@ packages:
   name: peilbeheerst-model
   version: 0.1.0
   path: src/peilbeheerst_model
-  sha256: 21b935f0542b5dbb76baaf1944b11ef6fab94b5a5926565410fe3d62722c332d
+  sha256: e8c270f68d683c802990a8bb905cd665fd327b33e1d17bf3e01dfe9c49ec335e
   requires_dist:
   - geopandas
+  - matplotlib
   - numpy
   - pandas
   - pydantic
