+++ src/peilbeheerst_model/peilbeheerst_model/__init__.py
@@ -1,6 +1,7 @@
 __version__ = "0.1.0"
 
 from peilbeheerst_model.parse_crossings import ParseCrossings
+from peilbeheerst_model.shortest_path import shortest_path_waterschap
 from peilbeheerst_model.waterschappen import waterschap_data
 
-__all__ = ["ParseCrossings", "waterschap_data"]
+__all__ = ["ParseCrossings", "shortest_path_waterschap", "waterschap_data"]
