+++ src/peilbeheerst_model/pyproject.toml
@@ -14,7 +14,17 @@ authors = [
 license = { text = "MIT" }
 
 requires-python = ">=3.10"
-dependencies = ["geopandas", "numpy", "pandas", "pydantic", "shapely", "tqdm", "matplotlib"]
+dependencies = [
+    "geopandas",
+    "numpy",
+    "pandas",
+    "pydantic",
+    "shapely",
+    "tqdm",
+    "matplotlib",
+    "fiona",
+    "networkx",
+]
 dynamic = ["version"]
 
 [project.optional-dependencies]
